Issue: #31 Edit: New directory structure
