+++ platform/analyzer/scripts/config.py
